added logging and the console so now maybe we can actually make some progress on interesting stuff

+++ catengine/Polygon.h
@@ -57,7 +57,8 @@ namespace catengine {
 
     void remove(_integer i)
     {
-      if (i < vertices_.size()) vertices_.erase(vertices_.begin() + i);
+      if (static_cast<size_t>(i) < vertices_.size()) 
+        vertices_.erase(vertices_.begin() + i);
     }
 
     void insert(_integer i, Vector2 const& vertex)
